Make trace be buffered
Rather then having the job trace shuffle the bytes over a channel to the
RunHandler so it can call out the the gitlab API make it simple append
to a buffer. This allows trace to become synchronous.

It's now the job of the JobHandler to push the trace buffer back to
gitlab very so often, currently simply every 3 seconds. Which more
matches the behaviour of the official gitlab runner.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>