Introduce an artifact cache with (optional) storage backing
Add a cache for (downloaded) artifacts to a job so they're only ever
downloaded at most once. Also store artifacts over 64K on storage rather
then in memory, even if the runner runs on a box with lots of memory the
glibc malloc doesn't behave great wrt. giving back memory. So store
bigger artifacts in storage, which could ofcourse be a tmpfs.