Add an API to insert completed jobs into the mock
In order to use the mock to test runners, we sometimes need to create
jobs to represent earlier steps in a gitlab pipeline. In particular, a
given runner might depend on artifacts from a previous step, if it
does not automatically clone the target repository itself, or if it
supports dynamically generated inputs.