Add support for cancellation
This adds a new parameter to JobHandler::step which can be queried to
determine whether the Gitlab job has been cancelled. It is then the
responsibility of the impl of JobHandler to determine how to
respond. Note that no further updates can be made to a Gitlab job once
it has been cancelled.

Behind the scenes, a monitor is spawned for each Gitlab job to query
whether the job has been cancelled. Communication is via a simple
flag: we call `make_canceler` which returns a pair `Canceler` and
`CancelClient`. The `Canceler` can be used by the monitor process to
signal that cancellation has occured; the `CancelClient` is handed to
the `JobHandler` impls so that they can check.