Add more of Gitlab's features to the enumeration
Note that we do not serialize them when they are not set, so there is
no observable change in terms of our interaction with Gitlab from this
commit.

There are a number of features that we may eventually want to enable,
and it's convenient to have a slightly wider range to pick from.