Stop passing the whole response to Uploader
This is in preparation for creating multiple uploaders, one for
each top level artifact that Gitlab is requesting. It doesn't make
sense to pass the data for all of them, so pull out the parts we
actually use.

Additionally, start passing the artifact we intend to upload, and
check that it's actually a type we can handle.