Support upload multiple artifacts
A lot of preparatory work has already been done before this step.
However, we still have to enable the feature flag, and we need to
switch to the more elaborate upload style that goes with it, where the
upload form now has additional metadata that Gitlab can use to
understand what is being provided.

There is also a change to the name of the artifacts, to match the
upstream runner: that provides foo.gz and foo.zip for the name 'foo',
depending on the compression model, and has a default basename of
'default' if none is provided.

The upstream MR that introduced this feature is particularly helpful:

https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/923/diffs#e7eb3b5ba14a314996a9e5fbbc6ffd18d6f53cc3