Add a test for uploading multiple artifacts
This adds a new similar, but separate, test into the artifact
integration test to check the behaviour when uploading multiple
artifacts. It specifically checks things like whether the default
archive name is set correctly, whether the artifact types are being
set correctly, and whether the file contents are correct.

The only wrinkle here is that because it joined an existing test in
the same file, tokio will attempt to parallelise it with the old test,
and that means we need to set the default subscriber for tracing more
carefully (as setting it globally does not work with two tests running
in parallel).