Add automatic generation of system_id for the runner
In Recent gitlab the runners are identified not only by token but also a
system_id. This allows gitlab to differantiate between different
instances of a single runner.

Gitlab allows the system_id to be an arbitrary text strings
of up to 64 characters. However the gitlab runner implementation creates
an id that's either prefixed by s_ (for machine-id based ids) or r_ (for
random ids) followed by 12 characters with the unique id. This
implements the same algorithm to create a system_id for consistency. As
opposed to gitlab-runner no persistant storage is used to keep the crate
stateless.