Add a per-layer filter on top of the GitlabLayer
tracing_subscriber supports per-layer filtering. This partially
works by the Filtered wrapper having a private thread-local variable
as a backchannel to the Registry. So to have functional per-layer
filtering the Filtered wrapper is required.

Adjust the GitlabLayer constructor to actually create one wrapped in a
Filtered, with a basic filter to only enable spans and events which have
a gitlab related field

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>