Seperate logging layer creation from runner creation
The runner crate uses tracing to log to gitlab, for this a tracing layer
is provided which should be part of the global subscriber. However
tracing makes it tricky to change the global subscriber at runtime, it's
somewhat assumed to be initialized early once.

Which makes runner creation also doing the layer creation slightly
awkward as it makes it hard to start logging before or during layer
creation.

Switch this around by letting the user first create the layer to output
to gitlab and associated job list. As the layer is independant of any
configuration (it just glues tracing to jobs), this can be done very
early to start logging asap.

In principle this also makes it possible for one process to create
multiple runners, though the assumption is made job identifiers are
unique, which fundamentally constrains all runners to a single gitlab
server.

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>