actions: Switch clippy to a dedicated action
Currently the "clippy latest" action starts failing whenever there are
new lints in the latest clippy giving warnings.  The main reason for the
latest lint is really to provide early warnings, rather then to act as a
pass/fail.

Switch to the rs-clippy-check action which wraps clippy and creates
annotations on issues, such that they're visibile in the review. For
clippy latest remove `-D warnings` which avoids the lint failing on
warnings. For the clippy stable check, warnings stay denied as that's
meant to provide a pass/fail result